+++ jax_moseq/utils/utils.py
@@ -4,8 +4,6 @@ import jax.numpy as jnp
 from sklearn.decomposition import PCA
 from jax.scipy.linalg import cho_factor, cho_solve
 from textwrap import fill
-from jax_moseq.utils.autoregression import ar_log_likelihood
-from functools import partial
 
 def symmetrize(A):
     """Symmetrize a matrix."""
@@ -209,26 +207,3 @@ def batch(data_dict, keys=None, seg_length=None, seg_overlap=30):
     mask = np.stack(mask)
     return stack,mask,labels
 
-def state_cross_likelihoods(params, states, mask, **kwargs):
-    """
-    Calculate log likelihoods of frames assigned to each state,
-    given the dynamics of each other state. See page 33 of the
-    supplement (Wiltchsko, 2015) for a formal definition.
-    """
-    x, Ab, Q = jax.device_put((states['x'],params['Ab'],params['Q']))
-    log_likelihoods = jax.lax.map(partial(ar_log_likelihood, x), (Ab, Q))
-    
-    nlags = mask.shape[1] - log_likelihoods.shape[2]
-    log_likelihoods = np.moveaxis(log_likelihoods,0,2)[mask[:,nlags:]>0]
-
-    z = states['z'][mask[:,nlags:]>0]
-    changepoints = np.diff(z).nonzero()[0]+1
-    counts = np.bincount(z[changepoints])
-    
-    n_states = log_likelihoods.shape[1]
-    cross_likelihoods = np.zeros((n_states,n_states))
-    for j in range(n_states):
-        ll = log_likelihoods[z==j].sum(0)
-        cross_likelihoods[j] = (ll-ll[j])/(counts[j] + 1e-6)
-    return cross_likelihoods
-
